Resolve T101 warnings
Mostly, I removed the comments but, in a couple places, I added a `# noqa: T101'
comment to let `flake8' know those TODO comments are okay.

I also created a discussion (#149) for us to track whether we want to leave
blank lines between sequence diagrams.

Additionally, #125 resolves most of the comments that were in `validator.py'
although one of the clean up comments, I did actually resolve in the code.

As far as the `test_parser.py' comment goes, I'll work on the tests in a future
commit.