More tests
Add initialize and initialized handlers to the server.

Catch failures when starting the server, waiting for it to become ready, etc.

Try doing things asynchronously - i.e. maybe we're not waiting long enough for
VSCode to send send the request. If this isn't true, though, then I think the
problem is with pygls not doing anything with the request.

Update the hover provider function.