Implement LSP server (#206)
* Add language server command

* Start the TCP server in a thread

* Remove threading

* Update server.pygls

Get rid of custom event loop. It seems like we can use `logging' to record the
requests.

Register a simple handler for `textDocument/hover' events.

* Add a simple (temporary) module to test the server

* Update server.py and test_server.py

Switch to using the standard `LanguageServer' since we're not trying to
customize the Event Loop, etc anymore and (at least for now) we don't have any
custom commands.

Try adding a feature for `initialize' and `initialized'.

Clean up the test client.

* Add vscode-langaugeclient as a dependency

* Update launch configs

* Add language configurations and property configurations

* Allow users to configure the path to python

The python extension configures this, but we can't use that because we're not in
the python extension and end users may not have that extension installed.

* Update extension to check preconditions

Things like making sure we have the correct python and aac versions installed.

* Remove test_server.py

It's causing dependency failures and we're not currently using it, anyways.

* Fix TS linting issues

* Update the VSCode extension

Start the language server from the VSCode extension.

* Remove custom initialize and initialized methods from the server

* Add basic tests for server.py

* Extract LSP-related functionality into a separate module

* Minor clean up

* Update server.py

Fix server logging (partially) in the sense that now we can at least see the
"Starting server..." message in the logs.

Return a hard-coded Hover object for testing.

* Don't re-start the server if it's already running

* Minor changes and try adding verbose logging

* Register a client-side hover provider

Register a hover provider on the client side to send the hover request to the
server - this isn't working either...

* Add missing import

Forgot to add this to my last commit.

* Swap order of parameters to ensureTrue

* More tests

Add initialize and initialized handlers to the server.

Catch failures when starting the server, waiting for it to become ready, etc.

Try doing things asynchronously - i.e. maybe we're not waiting long enough for
VSCode to send send the request. If this isn't true, though, then I think the
problem is with pygls not doing anything with the request.

Update the hover provider function.

* Ignore TypeScript server logs

* Update flake8

* Add configuration items for setting the debugHost and debugPort

* Add a --dev flag to the start-lsp sub-command

This flag allows specifying which server to use, the development server (i.e.
the WebSockets server) or the production server (i.e. the I/O server).

* Update AacLanguageServer.ts

Update the extension to start a different server depending on whether we're in
development or production mode.

* Use the TCP server instead of the WebSockets server for development

* Fix TS linter warning

* Move counter/ models to python/model/counter/

* Increase code coverage for server.py

Currently, it doesn't really make sense to test the handle_initialize and
handle_hover functions because they are returning static content. Once they are
returning something more realistic, we can add some tests for that.

* Minor fixes

* Add a basic hard-coded completion handler

* Add a bit of smarts to completion

* Minor adjustments

* Update server.py and add more tests

Delay creating the LanguageServer until the user requests running it and add
tests that exercise the hover and completion handlers.

* Fix linter warnings

* Update how we configure the logger

* Remove unused TEST_DOCUMENT variable

* Remove references to the TCP server

* Simplify sending requests to the server in tests

* Remove remaining references to host/port

* Hover provider is not needed, apparently

* Extract try-catch into a higher-level function

* Update ensure and ensureTrue

Pass the test, first, since it reads a bit more clearly, that way.

* Rename messageIfTestFails -> failureMessage

* Update debug launch config for the LSP server command

* Update docstring: Arguments -> Args

* Fix LSP server launch configuration name

* Increase code coverage

Since the newly added server.py has full code coverage, update test_gen_json.py
to use the new temporary_test_file context manager and add a test to check the
parser result for an unparsable file.

* Increase coverage threshold

* Fix linter warnings

* Move execShell per review comments

Put execShell into helpers.ts.

Rename maybeConvertBufferToString -> convertBufferToString.

* Remove unused window import

* Add a TSDoc comment to execShell

* Update dependencies and eslint configuration to check TSDoc comments

* Rename makeLspClient -> getLspClient

* Switch from using ensure* functions to assert.ok

* Rename getProdServerOptions -> getServerOptions

* Revert "Switch from using ensure* functions to assert.ok"

This reverts commit 3b061a8b55c6e665de1d4c843b88fb60bc8138e0.

Use ensureTrue again since assert.ok doesn't result in messages the user can see
and take action on.

* Rename configure_test.py -> LspTestClient.py

* Add documentation to LspTestClient and it's methods

Also, I extracted the common functionality for creating a server and server
thread into a single, more generic function.

* We don't need this

* Add return type to function signature and formatting

* Apply suggestions from code review

Co-authored-by: Alex Monnet <60155541+Coffee2Bits@users.noreply.github.com>

* Update python/tests/lang/LspTestClient.py

* Change ensure functions to be assert functions

* Switch from throwing errors to showing error messages

This will actually let the user know that something went wrong.

* Handle the case where Python 2 is being used

Apparently Python 2 writes its version output to standard error, so handle that
correctly, too.

* Update coverage threshold

I don't know where the rest of the coverage went, but apparently we're only at
91.66, right now.

Co-authored-by: Alex Monnet <60155541+Coffee2Bits@users.noreply.github.com>